Animation addition and filename change
Added animation to initial page to allow text to slide down
to the bottom of the page and remain there. This was done through
CSS animations by the creation of a keyframe rule I called
"slideDown".  Additionally, done by the addition of the
"animation-fill-mode"  rule which was set to keep the animated
content from going back to its initial state when the
animation had completed.

Also, filenames "index.js" and "index.css" were changed to "script.js"
and "style.css" respectively.

"Alert" method was also removed as it was only a test to ensure the
javascript file was loading correctly and ran as expected.